AoC_2022/pyproject.toml
[tool.poetry]
name = "aoc_2022"
version = "0.1.0"
description = ""
authors = ["Vojta Kuban <xmalmed@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "*"
parse = "*"
python-dotenv = "*"
requests = "*"
pytest = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
